Split out AI Chat Lab types
