+++ apps/src/aichat/redux/thunks/initializeAiCustomizations.ts
@@ -1,4 +1,9 @@
 import {
+  AiCustomizations,
+  LevelAichatSettings,
+  Visibility,
+} from '@cdo/apps/aichatLab/types';
+import {
   DEFAULT_VISIBILITIES,
   EMPTY_AI_CUSTOMIZATIONS,
 } from '@cdo/apps/aichatLab/views/modelCustomization/constants';
@@ -7,7 +12,6 @@ import {EVENTS} from '@cdo/apps/metrics/AnalyticsConstants';
 import {getTypedKeys} from '@cdo/apps/types/utils';
 import {AppDispatch} from '@cdo/apps/util/reduxHooks';
 
-import {AiCustomizations, LevelAichatSettings, Visibility} from '../../types';
 import {setInitialConfiguration} from '../slice';
 
 import {sendAnalytics} from './sendAnalytics';
