+++ apps/src/aichat/types/index.ts
@@ -3,8 +3,8 @@ export * from './analytics';
 export * from './assets';
 export * from './chatComponents';
 export * from './chatEvents';
-export * from './customizations';
-export * from './levelProperties';
+export * from './context';
+export * from './model';
 export * from './structuredOutput';
 export * from './systemPrompt';
 export * from './toxicity';
