+++ apps/src/aichatLab/types/levelProperties.ts
