+++ apps/src/aichatLab/views/modelCustomization/ExampleTopicsInputs.tsx
@@ -2,7 +2,7 @@ import Alert, {alertTypes} from '@code-dot-org/component-library/alert';
 import React, {useCallback, useMemo} from 'react';
 
 import {setModelCardProperty} from '@cdo/apps/aichat/redux';
-import {Visibility} from '@cdo/apps/aichat/types';
+import {Visibility} from '@cdo/apps/aichatLab/types';
 import {useAppDispatch} from '@cdo/apps/util/reduxHooks';
 
 import MultiInputCustomization from './MultiInputCustomization';
