+++ apps/src/aichatLab/views/modelCustomization/utils.ts
@@ -1,5 +1,5 @@
 import {modelDescriptions} from '@cdo/apps/aichat/constants';
-import {Visibility} from '@cdo/apps/aichat/types';
+import {Visibility} from '@cdo/apps/aichatLab/types';
 import Lab2Registry from '@cdo/apps/lab2/Lab2Registry';
 import {ValueOf} from '@cdo/apps/types/utils';
 import {AiChatModelIds} from '@cdo/generated-scripts/sharedConstants';
