+++ apps/src/lab2/levelEditors/aichatSettings/FieldSection.tsx
@@ -2,7 +2,7 @@ import {Typography} from '@mui/material';
 import classNames from 'classnames';
 import React, {useContext} from 'react';
 
-import {AiCustomizations} from '@cdo/apps/aichat/types';
+import {AiCustomizations} from '@cdo/apps/aichatLab/types';
 
 import {UpdateContext} from './UpdateContext';
 import VisibilityDropdown from './VisibilityDropdown';
