+++ apps/src/lab2/levelEditors/aichatSettings/VisibilityDropdown.tsx
@@ -1,7 +1,7 @@
 import {SimpleDropdown} from '@code-dot-org/component-library/dropdown';
 import React, {useContext} from 'react';
 
-import {AiCustomizations, Visibility} from '@cdo/apps/aichat/types';
+import {AiCustomizations, Visibility} from '@cdo/apps/aichatLab/types';
 
 import {UpdateContext} from './UpdateContext';
